Add optimisation for ticks in whammy
It often turns out to be the case that the optimal subpath from ticks T
and T + 1 are identical, in which case it makes sense to try the best
subpaths from tick T for tick T + 1 to see if one can skip the normal
optimisation procedure. We now implement this optimisation.